+++ src/features/maintenance/schemas/maintenance-ticket.schema.ts
@@ -99,11 +99,15 @@ export class MaintenanceTicket extends Document implements SoftDelete {
   @Prop({ maxlength: 1000, required: false })
   refuseReason?: string;
 
-  @Prop({ type: Number, required: false })
-  cost?: number;
-
   @Prop({ required: true, unique: true })
   ticketNumber: string;
+
+  @Prop({
+    type: MongooseSchema.Types.ObjectId,
+    ref: 'ScopeOfWork',
+    required: false,
+  })
+  scopeOfWork?: Types.ObjectId;
 
   deleted: boolean;
   deletedAt?: Date;